Moved `ReadOnlyBlackboardView` from package `de.uka.ipd.sdq.beagle.analysis` to package `de.uka.ipd.sdq.beagle.core`.
